requirements.txt
multitasking==0.0.11
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mailman==1.1.1
Flask-WTF==1.2.1
Flask-SocketIO==5.3.6
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
WTForms==3.1.2
Werkzeug==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
yfinance==0.2.28
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
stripe==10.5.0
psycopg2-binary==2.9.9
redis==5.0.7
scikit-learn==1.3.0
scipy==1.10.1
joblib==1.3.2
aiohttp==3.9.5
websockets==12.0
lxml==5.2.2
feedparser==6.0.10
APScheduler==3.10.4
beautifulsoup4==4.12.3
python-dateutil==2.9.0
pytz==2024.1
email-validator==2.1.0
python-jose==3.3.0
PyJWT==2.8.0
cryptography==41.0.7
celery==5.4.0
eventlet==0.33.3
psutil==5.9.8
# ta-lib==0.4.28  # Kommentert ut pga. Railway build issues. Installeres kun hvis nødvendig og C-bibliotek er tilgjengelig.
